Only use `:hover` pseudo-class
Reason: Although this _kind of_ break keyboard navigation,
I'm taking it as a trade-off. Previously, the element would
remain visible until **anything** else on the page was
focused; that isn't intended, but of course the rule was doing
what I told it to do. However, that seems rather _unpolished_

Giving up on a11y for the sake of _visuals_ is not uncommon.

Right?